removed unneeded files for aws connections
